add database entity for weather data, cleanup in JSON mapping, add mapper for saving weather data as database entities+unit test for it

+++ src/main/java/com/example/googlechartsthymeleaf/json_model/Wind.java
@@ -21,10 +21,10 @@ import lombok.Setter;
 public class Wind {
 
     @JsonProperty("speed")
-    public Double speed;
+    public Float speed;
     @JsonProperty("deg")
-    public Integer deg;
+    public Short deg;
     @JsonProperty("gust")
-    public Double gust;
+    public Float gust;
 
 }
